新增内容

+++ 算法专项/高精度/阶乘和.cpp
@@ -3,7 +3,7 @@ using namespace std;
 int main(){
     int n;
     cin>>n;
-    int ans=1,sum=0;
+    long long ans=1,sum=0;
     for(int i=1;i<=n;i++){
         ans*=i;
         sum+=ans;
